kernel/smp: Provide CSD lock timeout diagnostics
This not-for-mainline commit causes csd_lock_wait() to emit
diagnostics when a CPU fails to respond quickly enough to one of the
smp_call_function() family of function calls.  These diagnostics
include NMI stack traces, and so the exclusion of idle CPUs is
also removed.

This commit was inspired by an earlier patch by Josef Bacik.

Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
[ paulmck: Avoid 64-bit divides per kernel test robot feedback. ]
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>