percpu: fix first chunk size calculation for populated bitmap
Variable populated, which is a member of struct pcpu_chunk, is used as a
unit of size of unsigned long.
However, size of populated is miscounted. So, I fix this minor part.

The first chunk is special as it holds a preconfigured number of dynamic
pages (default 7). However, the first chunk size calculation only added
a single byte (instead of # of longs) to hold the populated bitmap. I
don't believe this was a widespread issue because: first, it works for
the default case and second, the dynamic page configuration is very
rarely tuned.

Fixes: 8ab16c43ea79 ("percpu: change the number of pages marked in the first_chunk pop bitmap")
CC: <stable@vger.kernel.org> # 4.14+
Signed-off-by: Sunghyun Jin <mcsmonk@gmail.com>
Signed-off-by: Dennis Zhou <dennis@kernel.org>
[Added explanation why I think I haven't seen a reported issue.]