mm: workingset: ignore slab memory size when calculating shadows pressure
In the memcg case count_shadow_nodes() sums the number of pages in lru
lists and the amount of slab memory (reclaimable and non-reclaimable) as a
baseline for the allowed number of shadow entries.

It seems to be a good analogy for the !memcg case, where
node_present_pages() is used.  However, it's not quite true, as there two
problems:

1) Due to slab reparenting introduced by commit fb2f2b0adb98 ("mm:
   memcg/slab: reparent memcg kmem_caches on cgroup removal") local
   per-lruvec slab counters might be inaccurate on non-leaf levels.  It's
   the only place where local slab counters are used.

2) Shadow nodes by themselves are backed by slabs.  So there is a loop
   dependency: the more shadow entries are there, the less pressure the
   kernel applies to reclaim them.

Fortunately, there is a simple way to solve both problems: slab counters
shouldn't be taken into the account by count_shadow_nodes().

Link: https://lkml.kernel.org/r/20200903230055.1245058-1-guro@fb.com
Signed-off-by: Roman Gushchin <guro@fb.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Shakeel Butt <shakeelb@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>