mm/vmscan: add a fatal signals check in drop_slab_node
On our server, there are about 10k memcg in one machine.  They use memory
very frequently.  We have observed that drop_caches can take a
considerable amount of time, and can't stop it.

There are two reasons:
1. There is somebody constantly generating more objects to reclaim
  on drop_caches, result the 'freed' always bigger than 10.
2. The process has no chance to process signals.

We can get the following info through 'ps':

 root:~# ps -aux | grep drop
 root  357956 ... R    Aug25 21119854:55 echo 3 > /proc/sys/vm/drop_caches
 root 1771385 ... R    Aug16 21146421:17 echo 3 > /proc/sys/vm/drop_caches

Add a bail out on the fatal signals in the main loop so that the
operation can be terminated by userspace.

Link: https://lkml.kernel.org/r/20200916025359.70203-1-zangchunxin@bytedance.com
Signed-off-by: Chunxin Zang <zangchunxin@bytedance.com>
Signed-off-by: Muchun Song <songmuchun@bytedance.com>
Acked-by: Michal Hocko <mhocko@suse.com>
Acked-by: Chris Down <chris@chrisdown.name>
Cc: David Hildenbrand <david@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>